split up modules to isolate Gtk dependencies
